pyproject.toml
[tool.poetry]
name = "mem0-exploration"
version = "0.1.0"
description = ""
authors = ["Leo Kim <leonardkim96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.57.2"
mem0ai = "^0.1.34"
ruff = "^0.8.2"
black = "^24.10.0"
pytest = "^8.3.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
